Re-enabled third_party_cookie_phaseout flag Moved duplicated methods "set_adapter_shortcut" and "set_adapter_url" to common_steps.py Added 2 arguments to "set_adapter_url" to be compatible with all the adapters Changed adt_plus_scenarios.py to call adapter_login_steps.launch_adapter() instead of having the logic in the file Added a constant (NS_EXTENSION_NAME) to adapter_login_steps.py for the NetSuite extension name Fixed urls on qa01.yml for the adapters
