+++ step_definitions/common_steps.py
@@ -2,9 +2,11 @@ import common
 import driver
 from selenium.common import NoSuchElementException
 from selenium.webdriver.common.by import By
+from selenium.webdriver import ActionChains
 from page_objects.common_page import CommonPage
 from pytest_bdd import given, when, parsers
 from step_definitions import login_steps
+from selenium.webdriver import Keys
 
 STARTED_PAGES = []
 COMMON_PAGE = CommonPage()
@@ -16,6 +18,11 @@ MODAL_TYPES = {
     'manual': '(Manual)'
 }
 AGENT_CREDENTIALS = {}
+PLATFORM_HOTKEYS = {
+    'windows': [Keys.CONTROL, 'ctrl'],
+    'mac': [Keys.COMMAND, 'command'],
+    'linux': [Keys.CONTROL, 'control']
+}
 TEARDOWN = False
 
 
@@ -203,6 +210,81 @@ def select_modal_next_button():
                 pass
     common.system_wait(1)
     assert button_clicked, "COULD NOT FIND OR CLICK IN NEXT BUTTON"
+
+
+def set_adapter_shortcut(driver_, extension_text, return_id=False):
+    driver_.get("chrome://extensions")
+    extension_manager_element_shadow = driver_.find_element(By.XPATH, "//extensions-manager")
+    extension_items_list_shadow = extension_manager_element_shadow.shadow_root.find_element(By.CSS_SELECTOR,
+                                                                                            "#items-list")
+    extensions = extension_items_list_shadow.shadow_root.find_elements(By.CSS_SELECTOR, "extensions-item")
+    selected_extension = None
+    for extension in extensions:
+        extension_name = extension.shadow_root.find_element(By.CSS_SELECTOR, "#name").text.strip()
+        if extension_name == extension_text:
+            selected_extension = extension
+            break
+    if not selected_extension:
+        raise Exception(f"extension {extension_text} not found")
+
+    # open extension options
+    selected_extension.shadow_root.find_element(By.CSS_SELECTOR, "#detailsButton").click()
+    common.wait_page_to_be(driver_, "?id=")
+    if return_id:
+        return driver_.current_url.split("?id=")[1].strip()
+
+    left_panel_shadow = extension_manager_element_shadow.shadow_root.find_element(By.CSS_SELECTOR, "extensions-sidebar")
+    keyboard_shortcuts = left_panel_shadow.shadow_root.find_element(By.CSS_SELECTOR, "a#sectionsShortcuts")
+    keyboard_shortcuts.click()
+
+    keyboard_shortcuts_content_shadow = extension_manager_element_shadow.shadow_root.find_element(By.CSS_SELECTOR,
+                                                                                                  "extensions-keyboard-shortcuts")
+    extension_cards = keyboard_shortcuts_content_shadow.shadow_root.find_elements(By.CSS_SELECTOR, ".shortcut-card")
+    selected_extension_card = None
+    for extension_card in extension_cards:
+        if extension_text in extension_card.text:
+            selected_extension_card = extension_card
+            break
+    if not selected_extension_card:
+        raise Exception(f"card for {extension_text} extension not found")
+
+    action = ActionChains(driver_)
+    edit_shortcut_shadow = selected_extension_card.find_element(By.CSS_SELECTOR, "extensions-shortcut-input")
+    edit_shortcut_button = edit_shortcut_shadow.shadow_root.find_element(By.CSS_SELECTOR, "cr-icon-button#edit")
+    edit_shortcut_button.click()
+    edit_shortcut_shadow.shadow_root.find_element(By.CSS_SELECTOR, "cr-input").shadow_root.find_element(By.CSS_SELECTOR, "#input").click()
+    action.key_down(PLATFORM_HOTKEYS.get(driver_.caps.get('platformName'))[0])\
+        .send_keys('i')\
+        .key_up(PLATFORM_HOTKEYS.get(driver_.caps.get('platformName'))[0])\
+        .perform()
+    
+def set_adapter_url(driver_, extension_id, steps_obj, adpt_url='adapter_login_url'):
+    lab_config = f"configuration/lab/{common.get_config_file_section('config.yml', 'configuration').get('lab')}.yml"
+    adapter_url = common.get_config_file_section(lab_config, 'configuration').get(adpt_url)
+    if not adapter_url:
+        adapter_url = steps_obj.url
+    steps_obj.url = adapter_url
+    driver_.get(f"chrome-extension://{extension_id}/options.html")
+    common.system_wait(2)
+
+    # set extension url
+    if adpt_url != 'adapter_login_url':
+        driver_.find_element(By.XPATH, "//select[@id='domain-select']/option[@value='Custom']").click()
+    input_url = driver_.find_element(By.XPATH, "//input[@id='url'] | //input[@id='custom-host']")
+    input_url.clear()
+    input_url.click()
+    input_url.send_keys(adapter_url)
+    try:
+        assert input_url.get_attribute('value') == adapter_url
+    except AssertionError:
+        input_url.clear()
+        input_url.send_keys(adapter_url)
+        assert input_url.get_attribute('value') == adapter_url, "COULD NOT SET THE CORRECT URL"
+    driver_.find_element(By.XPATH, "//button[@id='save']").click()
+    common.wait_element_attribute_contains(driver_, "//mark[@id='status']", 'innerText', "Options Saved")
+    tab_info = {'title': driver_.title,
+                'browser_number': int(common.get_driver_by_instance(driver_))}
+    common.BROWSER_TABS[driver_.current_window_handle] = tab_info
 
 
 @when("I check the new browser tab opened")
